requirements.txt
streamlit
torch
torchvision
transformers
sentence-transformers
faiss-cpu
scikit-learn
accelerate